+++ lib/src/sip_message.dart
@@ -2,7 +2,6 @@ import 'dart:convert' show utf8;
 
 import 'package:sdp_transform/sdp_transform.dart' as sdp_transform;
 
-import 'package:sip_ua/src/transactions/transaction_base.dart';
 import 'constants.dart' as DartSIP_C;
 import 'constants.dart';
 import 'data.dart';
@@ -10,7 +9,8 @@ import 'exceptions.dart' as Exceptions;
 import 'grammar.dart';
 import 'logger.dart';
 import 'name_addr_header.dart';
-import 'transport.dart';
+import 'socket_transport.dart';
+import 'transactions/transaction_base.dart';
 import 'ua.dart';
 import 'uri.dart';
 import 'utils.dart' as utils;
@@ -26,7 +26,10 @@ import 'utils.dart' as utils;
  * -param {String} [body]
  */
 class OutgoingRequest {
-  OutgoingRequest(this.method, this.ruri, this.ua, [Map<String, dynamic>? params, List<dynamic>? extraHeaders, String? body]) {
+  OutgoingRequest(this.method, this.ruri, this.ua,
+      [Map<String, dynamic>? params,
+      List<dynamic>? extraHeaders,
+      String? body]) {
     // Mandatory parameters check.
     if (method == null || ruri == null || ua == null) {
       throw Exceptions.TypeError('OutgoingRequest: ctor parameters invalid!');
@@ -35,7 +38,8 @@ class OutgoingRequest {
     params = params ?? <String, dynamic>{};
     // ignore: prefer_initializing_formals
     this.body = body;
-    if (extraHeaders != null) this.extraHeaders = utils.cloneArray(extraHeaders);
+    if (extraHeaders != null)
+      this.extraHeaders = utils.cloneArray(extraHeaders);
 
     // Fill the Common SIP Request Headers.
 
@@ -43,7 +47,9 @@ class OutgoingRequest {
     if (params['route_set'] != null) {
       setHeader('route', params['route_set']);
     } else if (ua.configuration.use_preloaded_route) {
-      setHeader('route', '<${ua.transport!.sip_uri};lr>');
+      if (ua.socketTransport != null) {
+        setHeader('route', '<${ua.socketTransport!.sip_uri};lr>');
+      }
     }
 
     // Via.
@@ -55,7 +61,9 @@ class OutgoingRequest {
 
     // To
     dynamic to_uri = params['to_uri'] ?? ruri;
-    dynamic to_params = params['to_tag'] != null ? <String, dynamic>{'tag': params['to_tag']} : null;
+    dynamic to_params = params['to_tag'] != null
+        ? <String, dynamic>{'tag': params['to_tag']}
+        : null;
     String? to_display_name = params['to_display_name'];
 
     to = NameAddrHeader(to_uri, to_display_name, to_params);
@@ -63,7 +71,9 @@ class OutgoingRequest {
 
     // From.
     dynamic from_uri = params['from_uri'] ?? ua.configuration.uri;
-    Map<String, dynamic> from_params = <String, dynamic>{'tag': params['from_tag'] ?? utils.newTag()};
+    Map<String, dynamic> from_params = <String, dynamic>{
+      'tag': params['from_tag'] ?? utils.newTag()
+    };
     String? display_name;
 
     if (params['from_display_name'] != null) {
@@ -78,7 +88,8 @@ class OutgoingRequest {
     setHeader('from', from.toString());
 
     // Call-ID.
-    String call_id = params['call_id'] ?? (ua.configuration.jssip_id! + utils.createRandomToken(15));
+    String call_id = params['call_id'] ??
+        (ua.configuration.jssip_id! + utils.createRandomToken(15));
 
     this.call_id = call_id;
     setHeader('call-id', call_id);
@@ -262,10 +273,8 @@ class OutgoingRequest {
     msg += 'User-Agent: $userAgent\r\n';
 
     if (body != null) {
-      logger.d('Outgoing Message: ${body!}');
-      //Here we should calculate the real content length for UTF8
-      List<int> encoded = utf8.encode(body!);
-      int length = encoded.length;
+      logger.d('Outgoing Message: $method body: $body');
+      int length = utf8.encode(body!).length;
       msg += 'Content-Length: $length\r\n\r\n';
       msg += body!;
     } else {
@@ -294,7 +303,9 @@ class OutgoingRequest {
 }
 
 class InitialOutgoingInviteRequest extends OutgoingRequest {
-  InitialOutgoingInviteRequest(URI? ruri, UA ua, [Map<String, dynamic>? params, List<dynamic>? extraHeaders, String? body]) : super(SipMethod.INVITE, ruri, ua, params, extraHeaders, body) {
+  InitialOutgoingInviteRequest(URI? ruri, UA ua,
+      [Map<String, dynamic>? params, List<dynamic>? extraHeaders, String? body])
+      : super(SipMethod.INVITE, ruri, ua, params, extraHeaders, body) {
     transaction = null;
   }
 
@@ -304,7 +315,8 @@ class InitialOutgoingInviteRequest extends OutgoingRequest {
 
   @override
   InitialOutgoingInviteRequest clone() {
-    InitialOutgoingInviteRequest request = InitialOutgoingInviteRequest(ruri, ua);
+    InitialOutgoingInviteRequest request =
+        InitialOutgoingInviteRequest(ruri, ua);
 
     headers.forEach((String? name, dynamic value) {
       request.headers[name] = List<dynamic>.from(headers[name]);
@@ -509,7 +521,7 @@ class IncomingRequest extends IncomingMessage {
   }
   UA? ua;
   URI? ruri;
-  Transport? transport;
+  SocketTransport? transport;
   TransactionBase? server_transaction;
   /**
   * Stateful reply.
@@ -520,7 +532,12 @@ class IncomingRequest extends IncomingMessage {
   * -param {Function} [onSuccess] onSuccess callback
   * -param {Function} [onFailure] onFailure callback
   */
-  void reply(int code, [String? reason, List<dynamic>? extraHeaders, String? body, Function? onSuccess, Function? onFailure]) {
+  void reply(int code,
+      [String? reason,
+      List<dynamic>? extraHeaders,
+      String? body,
+      Function? onSuccess,
+      Function? onFailure]) {
     List<dynamic> supported = <dynamic>[];
     dynamic to = getHeader('To');
 
@@ -609,6 +626,10 @@ class IncomingRequest extends IncomingMessage {
 
     if (body != null) {
       int length = body.length;
+      int utf8Length = utf8.encode(body).length;
+      if (length != utf8Length) {
+        logger.w('WARNING Non-ASCII character detected in message body');
+      }
 
       response += 'Content-Type: application/sdp\r\n';
       response += 'Content-Length: $length\r\n\r\n';
@@ -620,7 +641,8 @@ class IncomingRequest extends IncomingMessage {
     IncomingMessage message = IncomingMessage();
     message.data = response;
 
-    server_transaction?.receiveResponse(code, message, onSuccess as void Function()?, onFailure as void Function()?);
+    server_transaction?.receiveResponse(code, message,
+        onSuccess as void Function()?, onFailure as void Function()?);
   }
 
   /**
@@ -634,8 +656,6 @@ class IncomingRequest extends IncomingMessage {
     // Validate code and reason values.
     if (code == null || (code < 100 || code > 699)) {
       throw Exceptions.TypeError('Invalid status_code: $code');
-    } else if (reason != null) {
-      throw Exceptions.TypeError('Invalid reason_phrase: $reason');
     }
 
     reason = reason ?? DartSIP_C.REASON_PHRASE[code] ?? '';